Solution to Sort Students by Marks

+++ Sort Students by Marks@@67877d61dc107486ad5c2748/code.c
@@ -5,7 +5,7 @@ struct Student{
     char name[50];
     float marks;
 };
-void sort(int arr[],int n){
+void sort(float arr[],int n){
        for(int i=0;i<n;i++){
         for(int j=0;j<n-i-1;j++){
             if(arr[j]<arr[j+1]){
